Still not working but closer to executer

+++ rdna3emu/isa/memory.py
@@ -127,7 +127,8 @@ class Memory:
 
     # Untyped buffer store 2 dwords.
     # Values is an array of 2 dwords.
-    def global_store_b64(self, address, values: np.ndarray.astype(np.uint64)):
+    def global_store_b64(self, address, values: np.ndarray):
+        values = values.astype(np.uint64)
         # Sanity check the size of the array
         if len(values) != 2:
             raise Exception("Invalid array size")
@@ -135,7 +136,8 @@ class Memory:
             self.set_memory(address + i * 4, 4, values[i])
 
     # Untyped buffer store 3 dwords.
-    def global_store_b128(self, address, values: np.ndarray.astype(np.uint64)):
+    def global_store_b128(self, address, values: np.ndarray):
+        values = values.astype(np.uint64)
         # Sanity check the size of the array
         if len(values) != 3:
             raise Exception("Invalid array size")
@@ -143,7 +145,8 @@ class Memory:
             self.set_memory(address + i * 4, 4, values[i])
 
     # Untyped buffer store 4 dwords.
-    def global_store_b256(self, address, values: np.ndarray.astype(np.uint64)):
+    def global_store_b256(self, address, values: np.ndarray):
+        values = values.astype(np.uint64)
         # Sanity check the size of the array
         if len(values) != 4:
             raise Exception("Invalid array size")
